feat: comprehensive AE-dark theme + hardcoded color cleanup + lottie fix
- theme.rs: AE CC 2024 color palette, 7-tier backgrounds, 4-tier borders
- 20+ viewport overlay color constants (GRID_LINE, MOTION_PATH, GIZMO_X/Y/Z, etc.)
- panel_frame() and side_panel_frame() helper functions
- resize_corner_size increased to 8px
- Font sizes optimized (10px/11.5px/13px)
- 13 UI files cleaned of hardcoded colors → theme constants
- lottie_exporter: fix r## for hex colors, fix parse_anim_value_v2 flat/nested s format, clippy fix
- 380 tests pass, 0 clippy warnings

+++ src/core/tracker_engine.rs
@@ -441,7 +441,7 @@ mod subpixel_tests {
 #[cfg(test)]
 mod pixel_tracking_tests {
     use super::*;
-    use crate::core::timeline::{Composition, Layer, LayerType, TrackerPoint};
+    use crate::core::timeline::{Layer, LayerType, TrackerPoint};
     use crate::core::property::Animatable;
 
     /// Builds a 64x64 gray frame with a bright square at `sq_x`.
@@ -451,7 +451,7 @@ mod pixel_tracking_tests {
         let mut px = vec![30u8; w * h * 4];
         for y in 20..40 {
             for x in sq_x..sq_x + 12 {
-                let idx = ((y * w + x) * 4) as usize;
+                let idx = (y * w + x) * 4;
                 px[idx] = 240;
                 px[idx + 1] = 240;
                 px[idx + 2] = 240;
